Fixed the parameters on Client and Server startup. Tested that the Client-instance can connect to the server-instance of the game when running 2 instances of the game

+++ DataTransfer/server.cpp
@@ -9,7 +9,7 @@ Server::Server(QObject *parent) : QTcpServer(parent), isstatusok(false)
 
 Server::Server(QString port, QObject *parent) : QTcpServer(parent), isstatusok(false)
 {
-    qDebug() << "Creating server, port will be " << port;
+    qDebug() << "Creating server, port will be " << port.toInt();
     //serverstatus = Initialising;
 
     connect(this, SIGNAL(newConnection()), this, SLOT(onNewConnection()));
